+++ utils/Color.qml
@@ -1,0 +1,10 @@
+pragma Singleton
+
+import QtQuick
+import Quickshell
+
+Singleton {
+    function setAlpha(color, alpha) {
+        return Qt.rgba(color.r, color.b, color.a, alpha);
+    }
+}